Moving the timestamp functions to the API
They really shouldn't live in ImageAnalysis, and I'm not even sure how it was running because GEECS Scanner does not explicitly depend on ImageAnalysis.  Made a new folder in `tools` for these types of file operations.
